\name{airconditioner}
\alias{airconditioner}
\title{Electricity Usage in Air Conditioning Systems}
\usage{data(airconditioner)}
\description{
The \code{airconditioner} data frame has 20 observations on 3 
variables related to measurements on electricity usage during
a summer month for four different kinds of air conditioning 
systems.  The measurements were taken in houses that were 
randomly selected from five different home types which depended
on factors such as floor space, etc. 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{HomeType}{a factor representing type of home}
    \item{SystemType}{a factor representing the air conditioning system}
    \item{Usage}{a numeric vector representing electricity usage in KWh}
  }
}
\source{
 Devore, J.L., and Farnum, N. (2005)
Applied Statistics for Engineers and Scientists. 2nd Edition, Thomson.
}
\keyword{datasets}
