\name{airplane}
\alias{airplane}
\docType{data}
\title{Paper Airplane Flying Distances}
\description{
Flight distances (in meters) for 12 paper airplanes of varying weights.
}
\usage{data("airplane")}
\format{
  A data frame with 12 observations on 2 variables.
  \describe{
\item{\code{weight}}{factor with 3 levels}
\item{\code{distance}}{numeric flight distances}
  }
}
\keyword{datasets}
