\name{airplane.sim11}
\alias{airplane.sim11}
\docType{data}
\title{Simulated Paper Airplane Flying Distances - Replicate 3}
\description{
Simulated flight distances (in meters) for 12 paper airplanes of varying weights.
These data were generated under the assumption that there are differences
in mean flight difference due to differences in the weight of the paper.
The noise variance was assumed to be 0.96.  
}
\usage{data("airplane.sim01")}
\format{
  A data frame with 12 observations on 2 variables.
  \describe{
\item{\code{weight}}{factor with 3 levels}
\item{\code{distance}}{numeric flight distances}
  }
}
\keyword{datasets}
