\name{airplane2}
\alias{airplane2}
\docType{data}
\title{Paper Airplane Flying Distances Replicated Study}
\description{
Flight distances (in meters) for 20 paper airplanes of varying weights.
}
\usage{data("airplane2")}
\format{
  A data frame with 20 observations on 2 variables.
  \describe{
\item{\code{weight}}{factor with 4 levels}
\item{\code{distance}}{numeric flight distances}
  }
}
\keyword{datasets}
