\name{cigbutts}
\alias{cigbutts}
\docType{data}
\title{Cigarette Butts}
\description{
On a university campus there are a number of areas designated for 
smoking. Outside of those areas, smoking is not permitted. One of the 
smoking areas is towards the north end of the campus near some parking 
lots and a large walkway towards one of the residences. Along the 
walkway, cigarette butts are visible in the nearby grass. Numbers of 
cigarette butts were counted at various distances from the smoking area 
in 200x80 square-cm quadrats located just west of the walkway.}

\usage{data("cigbutts")}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{distance}}{distance from gazebo}
    \item{\code{count}}{observed number of butts}
  }
}
\keyword{datasets}
