\name{ClothStrength}
\alias{ClothStrength}
\title{Cloth Strength Measurements}
\usage{ClothStrength}
\description{
Strength measurements of 5 bolts of cloth, each treated with varying
amounts of a chemical.  
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Bolt}{a factor with 5 levels}
    \item{Chemical}{a factor with 4 levels}
    \item{Strength}{a numeric vector}
  }
}
\keyword{datasets}
