\name{motor}
\alias{motor}
\docType{data}
\title{Motor Vibration Data}
\description{
Noise measurements for 5 samples of motors, each sample based on
a different brand of bearing.
}
\usage{data("motor")}
\format{
  A data frame with 5 columns.
  \describe{
    \item{\code{Brand 1}}{A numeric vector length 6}
    \item{\code{Brand 2}}{A numeric vector length 6}
    \item{\code{Brand 3}}{A numeric vector length 6}
    \item{\code{Brand 4}}{A numeric vector length 6}
    \item{\code{Brand 5}}{A numeric vector length 6}
  }
}
\source{Devore, J. and N. Farnum (2005) Applied Statistics for
Engineers and Scientists.  Thomson.
}
\keyword{datasets}
