\name{p12.16}
\alias{p12.16}
\title{Data Set for Problem 12-16}
\usage{data(p12.16)}
\description{
The \code{p12.16} data frame has 26 observations on 5 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Mixture}{numeric}
    \item{x1}{numeric}
    \item{x2}{numeric}
    \item{x3}{numeric}
    \item{y}{numeric}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\references{
Myers, R. Technometrics, vol. 6, no. 4, 343-356, 1964. 
}
\keyword{datasets}
