\name{p12.8}
\alias{p12.8}
\title{Data Set for Problem 12-8}
\usage{data(p12.8)}
\description{
The \code{p12.8} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p12.8)
}
\keyword{datasets}
