\name{p13.1}
\alias{p13.1}
\title{Data Set for Problem 13-1}
\usage{data(p13.1)}
\description{
The \code{p13.1} data frame has 25 observation on the
test-firing results for surface-to-air missiles.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{target speed (in Knots)}
    \item{y}{hit (=1) or miss (=0)}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.1)
}
\keyword{datasets}
