\name{p13.16}
\alias{p13.16}
\non_function{}
\title{Data Set for Problem 13-16}
\usage{data(p13.16)}
\description{
The \code{p13.16} data frame has 16 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{X1}{a numeric vector}
    \item{X2}{a numeric vector}
    \item{X3}{a numeric vector}
    \item{X4}{a numeric vector}
    \item{Y}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.16)
}
\keyword{datasets}
