\name{p13.2}
\alias{p13.2}
\non_function{}
\title{Data Set for Problem 13-2}
\usage{data(p13.2)}
\description{
The \code{p13.2} data frame has 20 observations on home ownership.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{family income}
    \item{y}{home ownership (1 = yes, 0 = no)}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.2)
}
\keyword{datasets}
