\name{p13.20}
\alias{p13.20}
\title{Data Set for Problem 13-20}
\usage{data(p13.20)}
\description{
The \code{p13.20} data frame has 30 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{yhat}{a numeric vector}
    \item{resdev}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.20)
}
\keyword{datasets}
