\name{p13.3}
\alias{p13.3}
\title{Data Set for Problem 13-3}
\usage{data(p13.3)}
\description{
The \code{p13.3} data frame has 10 observations on the
compressive strength of an alloy fastener used in
aircraft construction.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{load (in psi)}
    \item{n}{sample size}
    \item{r}{number failing}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.3)
}
\keyword{datasets}
