\name{p13.4}
\alias{p13.4}
\title{Data Set for Problem 13-4}
\usage{data(p13.4)}
\description{
The \code{p13.4} data frame has 11 observations on the
effectiveness of a price discount coupon on the
purchase of a two-litre beverage.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{discount}
    \item{n}{sample size}
    \item{r}{number redeemed}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.4)
}
\keyword{datasets}
