\name{p13.5}
\alias{p13.5}
\non_function{}
\title{Data Set for Problem 13-5}
\usage{data(p13.5)}
\description{
The \code{p13.5} data frame has 20 observations on
new automobile purchases.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{income}
    \item{x2}{age of oldest vehicle}
    \item{y}{new purchase less than 6 months later (1=yes, 0=no)}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.5)
}
\keyword{datasets}
