\name{p13.6}
\alias{p13.6}
\non_function{}
\title{Data Set for Problem 13-6}
\usage{data(p13.6)}
\description{
The \code{p13.6} data frame has 15 observations
on the number of failures of a particular type of valve
in a processing unit.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{valve}{type of valve}
    \item{numfail}{number of failures}
    \item{months}{months}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p13.6)
}
\keyword{datasets}
