\name{p13.7}
\alias{p13.7}
\title{Data Set for Problem 13-7}
\usage{data(p13.7)}
\description{
The \code{p13.7} data frame has 44 observations on the coal
mines of the Appalachian region of western Virginia.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{number of fractures in upper seams of coal mines}
    \item{x1}{inner burden thickness (in feet), 
shortest distance between  seam floor and the lower
seam}
   \item{x2}{percent extraction of the lower previously
mined seam}
    \item{x3}{lower seam height (in feet)}
    \item{x4}{time that the mine has been in operation (in years)}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\references{
Myers (1990)
}
\examples{
data(p13.7)
}
\keyword{datasets}
