\name{p14.1}
\alias{p14.1}
\non_function{}
\title{Data Set for Problem 14-1}
\usage{data(p14.1)}
\description{
The \code{p14.1} data frame has 15 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{a numeric vector}
    \item{y}{a numeric vector}
    \item{time}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p14.1)
}
\keyword{datasets}
