\name{p14.2}
\alias{p14.2}
\title{Data Set for Problem 14-2}
\usage{data(p14.2)}
\description{
The \code{p14.2} data frame has 18 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{t}{a numeric vector}
    \item{xt}{a numeric vector}
    \item{yt}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p14.2)
}
\keyword{datasets}
