\name{p15.4}
\alias{p15.4}
\non_function{}
\title{Data Set for Problem 15-4}
\usage{data(p15.4)}
\description{
The \code{p15.4} data frame has 40 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{y}{a numeric vector}
    \item{set}{a factor with levels}
    \item{set}{\code{e} }
    \item{set}{\code{p} }
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p15.4)
}
\keyword{datasets}
