\name{p2.9}
\alias{p2.9}
\title{Data Set for Problem 2-9}
\usage{data(p2.9)}
\description{
The \code{p2.9} data frame has 25 rows and 2 columns.  See
help on \code{softdrink} for details.  
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{a numeric vector}
    \item{x}{a numeric vector}
  }
}

\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p2.9)
}
\keyword{datasets}
