\name{p5.1}
\alias{p5.1}
\non_function{}
\title{Data Set for Problem 5-1}
\usage{data(p5.1)}
\description{
The \code{p5.1} data frame has 8 observations on the impact
of temperature on the viscosity of toluene-tetralin blends.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{temp}{temperature}
    \item{visc}{viscosity}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\references{
Byers and Williams (1987) Viscosities of Binary and Ternary Mixtures
of Polyaromatic Hydrocarbons. Journal of Chemical and Engineering
Data, 32, 349-354.
}
\examples{
data(p5.1)
plot(p5.1)
}
\keyword{datasets}
