\name{p5.2}
\alias{p5.2}
\title{Data Set for Problem 5-2}
\usage{data(p5.2)}
\description{
The \code{p5.2} data frame has 11 observations on the vapor
pressure of water for various temperatures.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{temp}{temperature (K)}
    \item{vapor}{vapor pressure (mm Hg)}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p5.2)
plot(p5.2)
}
\keyword{datasets}
