\name{p5.4}
\alias{p5.4}
\title{Data Set for Problem 5-4}
\usage{data(p5.4)}
\description{
The \code{p5.4} data frame has 8 observations on 2 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(p5.4)
plot(y ~ x, data=p5.4)
}
\keyword{datasets}
