\name{p8.16}
\alias{p8.16}
\title{Data Set for Problem 8-16}
\usage{data(p8.16)}
\description{
The \code{p8.16} data frame has 17 observations on 4 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Location}{numeric}
    \item{INHIBIT}{numeric}
    \item{UVB}{numeric}
    \item{SURFACE}{character}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\references{
Smith, R. C. et al., "Ozone depletion: Ultraviolet radiation and phytoplankton biology in Antartic waters,"
 Science, 255, 952-957, 1992.
}
\keyword{datasets}
