\name{pathoeg}
\alias{pathoeg}
\title{Pathological Example}
\usage{litters}
\description{
Artificial regression data which causes stepwise regression
with AIC to produce a highly non-parsimonious model.  The
true model used to simulate the data has only one real
predictor (x8).   
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
    \item{x4}{a numeric vector}
    \item{x5}{a numeric vector}
    \item{x6}{a numeric vector}
    \item{x7}{a numeric vector}
    \item{x8}{a numeric vector}
    \item{x9}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\keyword{datasets}
