\name{rectangles}
\alias{rectangles}
\title{Length Measurements on Rectangular Objects}
\description{
Observations of heights, widths and diagonal lengths of several
rectangular objects, such as books, photographs, and so on were measured. 
}
\usage{rectangles}
\format{
  A data frame with 40 observations on the following 4 variables.
  \describe{
\item{\code{h}}{numeric, heights in centimeters}
\item{\code{w}}{numeric, widths in centimeters}
\item{\code{d}}{numeric, diagonal lengths in centimeters}
\item{\code{index}}{numeric, sum of squares of heights and widths}
  }
}
\keyword{datasets}
