\name{softdrink}
\alias{softdrink}
\title{Softdrink Data}
\usage{data(softdrink)}
\description{
The \code{softdrink} data frame has 25 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{a numeric vector}
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(softdrink)
}
\keyword{datasets}
