\name{solar}
\alias{solar}
\non_function{}
\title{Solar Data}
\usage{data(solar)}
\description{
The \code{solar} data frame has 29 rows and 6 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{total.heat.flux}{a numeric vector}
    \item{insolation}{a numeric vector}
    \item{focal.pt.east}{a numeric vector}
    \item{focal.pt.south}{a numeric vector}
    \item{focal.pt.north}{a numeric vector}
    \item{time.of.day}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(solar)
}
\keyword{datasets}
