\name{stackloss}
\alias{stackloss}
\title{Stackloss Data}
\usage{data(stackloss)}
\description{
The \code{stackloss} data frame has 21 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{a numeric vector}
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}

\examples{
data(stackloss)
}
\keyword{datasets}
