\name{stain}
\alias{stain}
\title{Stain Removal Data}
\usage{data(stain)}
\description{
Data on an experiment to remove ketchup stains from white cotton
fabric by soaking the stained fabric in one of five substrates for
one hour.  Remaining stains were scored visually and subjectively
according to a 6-point scale (0 = completely clean, 5 = no change)
The \code{stain} data frame has 15 rows and 2 columns.  
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{treatment}{a factor}
    \item{response}{a numeric vector}
  }
}
\examples{
data(stain)
}
\keyword{datasets}
