\name{table.b12}
\alias{table.b12}
\title{Table B12}
\usage{data(table.b12)}
\description{
The \code{table.b12} data frame has 32 rows and 6 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{temp}{a numeric vector}
    \item{soaktime}{a numeric vector}
    \item{soakpct}{a numeric vector}
    \item{difftime}{a numeric vector}
    \item{diffpct}{a numeric vector}
    \item{pitch}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(table.b12)
}
\keyword{datasets}
