\name{table.b13}
\alias{table.b13}
\title{Table B13}
\usage{data(table.b13)}
\description{
The \code{table.b13} data frame has 40 observations on 7 variables concerning jet turbine engine thrust.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{a numeric vector representing thrust}
    \item{x1}{a numeric vector representing primary speed of rotation}
    \item{x2}{a numeric vector representing secondary speed of rotation}
    \item{x3}{a numeric vector representing fuel flow rate}
    \item{x4}{a numeric vector representing pressure}
    \item{x5}{a numeric vector representing exhaust temperature}
    \item{x6}{a numeric vector representing ambient temperature at time of test}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(table.b13)
}
\keyword{datasets}
