\name{table.b16}
\alias{table.b16}
\title{Table B16 - Life Expectancy Data}
\usage{data(table.b16)}
\description{
The \code{table.b16} data frame has 38 observations on 6 variables. Each observation
corresponds to an individual country. 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Country}{character vector}
    \item{LifeExp}{numeric vector, in years}
    \item{People.per.TV}{numeric vector}
    \item{People.per.Dr}{numeric vector}
    \item{LifeExpMale}{numeric vector, in years}
    \item{LifeExpFemale}{numeric vector, in years}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\keyword{datasets}
