\name{table.b17}
\alias{table.b17}
\title{Table B17 - Satisfaction Survey}
\usage{data(table.b17)}
\description{
The \code{table.b17} data frame has 25 observations on 5 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Satisfaction}{numeric vector}
    \item{Age}{numeric vector, in years}
    \item{Severity}{numeric vector}
    \item{Surgical.Medical}{numeric vector}
    \item{Anxiety}{numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\keyword{datasets}
