\name{table.b19}
\alias{table.b19}
\title{Table B19}
\usage{data(table.b19)}
\description{
The \code{table.b19} data frame has 32 observations on 11 variables. 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{numeric vector}
    \item{x1}{numeric vector}
    \item{x2}{numeric vector}
    \item{x3}{numeric vector}
    \item{x4}{numeric vector}
    \item{x5}{numeric vector}
    \item{x6}{numeric vector}
    \item{x7}{numeric vector}
    \item{x8}{numeric vector}
    \item{x9}{numeric vector}
    \item{x10}{numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\keyword{datasets}
