\name{table.b20}
\alias{table.b20}
\title{Table B20}
\usage{data(table.b20)}
\description{
The \code{table.b20} data frame has 18 observations on 6 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{numeric vector}
    \item{x2}{numeric vector}
    \item{x3}{numeric vector}
    \item{x4}{numeric vector}
    \item{x5}{numeric vector}
    \item{y}{numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\examples{
pairs(table.b20)
}
\keyword{datasets}
