\name{tree.sample}
\alias{tree.sample}
\docType{data}
\title{Sample of Loblolly Pine Data}
\description{
A random sample of observations taken from the 'Loblolly' data frame, one
per Seed.   
}
\usage{data("tree.sample")}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{height}}{tree heights (ft)}
    \item{\code{age}}{tree ages (yr)}
  }
}
\keyword{datasets}
