\name{wxNWO}
\alias{wxNWO}
\title{Weather Observations for Three Stations in Northwestern Ontario}
\description{
Daily observations taken from 2012 through 2021 on temperature, rain, snow and wind for Fort Frances,
Kenora and Dryden, Ontario.  
}
\usage{wxNWO}
\format{
  A data frame with 10959 observations on the following 31 variables.
  \describe{
\item{\code{Longitude}}{numeric}
\item{\code{Latitude}}{numeric}
\item{\code{Station.Name}}{character}
\item{\code{Climate.ID}}{numeric}
\item{\code{Date.Time}}{numeric}
\item{\code{Year}}{numeric}
\item{\code{Month}}{numeric}
\item{\code{Day}}{numeric}
\item{\code{Data.Quality}}{numeric}
\item{\code{Max.Temp}}{numeric}
\item{\code{Max.Temp.Flag}}{numeric}
\item{\code{Min.Temp}}{numeric}
\item{\code{Min.Temp.Flag}}{numeric}
\item{\code{Mean.Temp}}{numeric}
\item{\code{Mean.Temp.Flag}}{numeric}
\item{\code{Heat.Deg.Days}}{numeric}
\item{\code{Heat.Deg.Days.Flag}}{numeric}
\item{\code{Cool.Deg.Days}}{numeric}
\item{\code{Cool.Deg.Days.Flag}}{numeric}
\item{\code{Total.Rain}}{numeric}
\item{\code{Total.Rain.Flag}}{numeric}
\item{\code{Total.Snow}}{numeric}
\item{\code{Total.Snow.Flag}}{numeric}
\item{\code{Total.Precip}}{numeric}
\item{\code{Total.Precip.Flag}}{numeric}
\item{\code{Snow.on.Ground}}{numeric}
\item{\code{Snow.on.Ground.Flag}}{numeric}
\item{\code{Dir.of.Max.Gust}}{numeric}
\item{\code{Dir.of.Max.Gust.Flag}}{numeric}
\item{\code{Speed.of.Max.Gust}}{numeric}
\item{\code{Speed.of.Max.Gust.Flag}}{numeric}
  }
}
\source{Environment Canada
}
\keyword{datasets}
