\name{table.b16}
\alias{table.b16}
\title{Table B16 Data Set}
\usage{data(table.b16)}
\description{
The \code{table.b16} data frame has 38 observations on 6 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Country}{character}
    \item{LifeExp}{numeric}
    \item{People.per.TV}{numeric}
    \item{People.per.Dr}{numeric}
    \item{LifeExpMale}{numeric}
    \item{LifeExpFemale}{numeric}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\keyword{datasets}
