\name{table.b18}
\alias{table.b18}
\title{Table B18}
\usage{data(table.b18)}
\description{
The \code{table.b18} data frame has 16 observations on 9 variables.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{numeric vector}
    \item{x1}{numeric vector}
    \item{x2}{numeric vector}
    \item{x3}{numeric vector}
    \item{x4}{numeric vector}
    \item{x5}{numeric vector}
    \item{x6}{numeric vector}
    \item{x7}{numeric vector}
    \item{x8}{numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2021)
Introduction to Linear Regression Analysis. 6th Edition, John Wiley and Sons.
}
\examples{
pairs(table.b18)
}
\keyword{datasets}
