\name{table.b2}
\alias{table.b2}
\title{Table B2}
\usage{data(table.b2)}
\description{
The \code{table.b2} data frame has 29 rows and 6 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{a numeric vector}
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
    \item{x4}{a numeric vector}
    \item{x5}{a numeric vector}
  }
}
\source{
 Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis. 3rd Edition, John Wiley and Sons.
}
\examples{
data(table.b2)
}
\keyword{datasets}
